+++ lib/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod core;
 pub mod dag_creator;
+pub mod dag_scheduler;
+pub mod dag_set_scheduler;
 pub mod fixed_priority_scheduler;
 pub mod global_edf_scheduler;
 pub mod graph_extension;
